Make legacy explain the concrete base, drop the extra Legacy class
Match the convention used by every other operation (getDocument, listDocuments,
etc.): the legacy Databases endpoint is the concrete Get that the TablesDB,
DocumentsDB, and VectorsDB surfaces extend — there is no separate Legacy class.
The base was needlessly abstract with a parallel Legacy.php; fold the legacy
route/metadata into Get and register it directly.
