--- src/Appwrite/Platform/Modules/Databases/Http/Databases/Collections/Documents/Explain/Legacy.php
@@ -1,57 +1,0 @@
-<?php
-
-namespace Appwrite\Platform\Modules\Databases\Http\Databases\Collections\Documents\Explain;
-
-use Appwrite\SDK\AuthType;
-use Appwrite\SDK\ContentType;
-use Appwrite\SDK\Method;
-use Appwrite\SDK\Response as SDKResponse;
-use Utopia\Database\Database;
-use Utopia\Database\Validator\UID;
-use Utopia\Http\Adapter\Swoole\Response as SwooleResponse;
-use Utopia\Validator\ArrayList;
-use Utopia\Validator\Boolean;
-use Utopia\Validator\Text;
-
-class Legacy extends Get
-{
-    public static function getName(): string
-    {
-        return 'explainDocuments';
-    }
-
-    public function __construct()
-    {
-        $this
-            ->setHttpMethod(self::HTTP_REQUEST_METHOD_GET)
-            ->setHttpPath('/v1/databases/:databaseId/collections/:collectionId/documents/explain')
-            ->desc('Explain documents query plan')
-            ->groups(['api', 'database'])
-            ->label('scope', 'documents.read')
-            ->label('resourceType', RESOURCE_TYPE_DATABASES)
-            ->label('sdk', new Method(
-                namespace: $this->getSDKNamespace(),
-                group: $this->getSDKGroup(),
-                name: self::getName(),
-                description: '/docs/references/databases/explain-documents.md',
-                auth: [AuthType::ADMIN, AuthType::SESSION, AuthType::KEY, AuthType::JWT],
-                responses: [
-                    new SDKResponse(
-                        code: SwooleResponse::STATUS_CODE_OK,
-                        model: $this->getResponseModel(),
-                    ),
-                ],
-                contentType: ContentType::JSON,
-            ))
-            ->param('databaseId', '', fn (Database $dbForProject) => new UID($dbForProject->getAdapter()->getMaxUIDLength()), 'Database ID.', false, ['dbForProject'])
-            ->param('collectionId', '', fn (Database $dbForProject) => new UID($dbForProject->getAdapter()->getMaxUIDLength()), 'Collection ID.', false, ['dbForProject'])
-            ->param('queries', [], new ArrayList(new Text(APP_LIMIT_ARRAY_ELEMENT_SIZE), APP_LIMIT_ARRAY_PARAMS_SIZE), 'Array of query strings generated using the Query class provided by the SDK. Same shape as listDocuments.', true)
-            ->param('total', true, new Boolean(true), 'When true, the explain captures the COUNT(*) call listDocuments fires for the total field as a second entry. Mirrors listDocuments default behavior.', true)
-            ->inject('response')
-            ->inject('dbForProject')
-            ->inject('user')
-            ->inject('getDatabasesDB')
-            ->inject('authorization')
-            ->callback($this->action(...));
-    }
-}
